# openapi/spec.yaml
openapi: 3.1.0
info:
  title: OAuth Implementation Server
  description: |
    A comprehensive OAuth 2.0 / OpenID Connect implementation server for testing all OAuth flows and standards.

    Supports: Authorization Code + PKCE, Client Credentials, Device Code (RFC 8628),
    CIBA (Client-Initiated Backchannel Authentication), PAR (RFC 9126), DPoP (RFC 9449),
    Token Introspection (RFC 7662), Token Revocation (RFC 7009), Dynamic Client Registration (RFC 7591),
    Resource Indicators (RFC 8707), OIDC Discovery, JWKS, UserInfo, TOTP-based MFA,
    Consent Management, Custom Scopes, and login_hint/prompt/max_age parameters.
  version: 1.1.0
  contact:
    name: API Support
    url: https://github.com/bravo68web/oauth-impl
servers:
  - url: https://localhost:8080
    description: Local development server
...
paths:
...
  /oauth/logout:
    get:
      tags: [OIDC]
      summary: RP-Initiated Logout
      description: Ends the browser session. A valid id_token_hint logs out without a prompt. Otherwise the response is a confirmation page.
      parameters:
        - in: query
          name: id_token_hint
          schema: { type: string }
        - in: query
          name: client_id
          schema: { type: string }
        - in: query
          name: post_logout_redirect_uri
          schema: { type: string }
        - in: query
          name: state
          schema: { type: string }
      responses:
        "200":
          description: Logged out, or a confirmation page
        "302":
          description: Redirect to a registered post_logout_redirect_uri
    post:
      tags: [OIDC]
      summary: Confirm RP-Initiated Logout
      responses:
        "200":
          description: Logged out
...
  /oauth/register:
    post:
      tags: [OAuth 2.0]
      summary: Dynamic Client Registration (RFC 7591)
      description: |
        Register a new OAuth 2.0 client dynamically. Returns client credentials including
        `client_id` and `client_secret`. The client can then be used with any supported grant type.

        See [RFC 7591](https://datatracker.ietf.org/doc/html/rfc7591) for details.
      operationId: registerClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DCRRequest"
            example:
              client_name: "My Web App"
              redirect_uris: ["https://example.com/callback"]
              grant_types: ["authorization_code", "client_credentials"]
              scope: "openid profile email"
      responses:
        "201":
          description: Client registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DCRResponse"
              example:
                client_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                client_secret: "s3cr3t-k3y-h3r3-4nd-l0ng3r-th4n-th1s"
                client_id_issued_at: 1718000000
                client_secret_expires_at: 0
                redirect_uris: ["https://example.com/callback"]
                token_endpoint_auth_method: "client_secret_basic"
                grant_types: ["authorization_code", "client_credentials"]
                response_types: ["code"]
                client_name: "My Web App"
                scope: "openid profile email"
        "400":
          description: Invalid request (missing required fields or invalid redirect URI)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
              example:
                error: invalid_request
                error_description: "client_name is required"
  /oauth/authorize:
    get:
      tags: [OAuth 2.0]
      summary: Authorization endpoint
      description: |
        Initiates the OAuth 2.0 authorization flow. The user-agent is redirected to this endpoint
        to authenticate the user and obtain consent. On success, redirects back to the client
        with an authorization code.

        For testing, pass `user_id` to auto-approve without user interaction.

        Supports PKCE via `code_challenge` and `code_challenge_method` parameters.
      operationId: authorize
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: The client identifier obtained from registration
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum:
              [
                code,
                token,
                id_token,
                "code id_token",
                "code token",
                "code id_token token",
                none,
              ]
          description: |
            OAuth response type:
            - `code`: Authorization code flow (query parameter)
            - `token`: Implicit access token (fragment)
            - `id_token`: Implicit ID token (fragment)
            - `code id_token`: Hybrid - code + ID token (fragment)
            - `code token`: Hybrid - code + access token (fragment)
            - `code id_token token`: Hybrid - all three (fragment)
            - `none`: No tokens returned
          example: code
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
          description: Must match one of the registered redirect URIs
          example: "https://example.com/callback"
        - name: scope
          in: query
          schema:
            type: string
          description: Space-separated list of requested scopes
          example: "openid profile email"
        - name: state
          in: query
          schema:
            type: string
          description: Opaque value for CSRF protection (recommended)
          example: "abc123xyz"
        - name: code_challenge
          in: query
          schema:
            type: string
          description: PKCE code challenge (base64url-encoded SHA256 of code_verifier)
          example: "E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM"
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256, plain]
          description: PKCE challenge method (S256 recommended)
          example: S256
        - name: nonce
          in: query
          schema:
            type: string
          description: OpenID Connect nonce for replay protection
          example: "n-0S6_WzA2Mj"
        - name: prompt
          in: query
          schema:
            type: string
            enum: [none, login, consent, select_account]
          description: |
            OIDC prompt parameter:
            - `none`: No UI shown; fail if not already authenticated and consented
            - `login`: Force re-authentication even if already logged in
            - `consent`: Force consent screen even if previously consented
            - `select_account`: Account chooser (not yet implemented)
          example: consent
        - name: login_hint
          in: query
          schema:
            type: string
          description: Hint to pre-fill the username on the login page
          example: "johndoe"
        - name: resource
          in: query
          schema:
            type: string
          description: |
            Resource indicator (RFC 8707). The URI of the resource server the client wants to access.
            The issued token will be scoped to this resource.
          example: "https://api.example.com"
        - name: user_id
          in: query
          schema:
            type: string
          description: User ID for auto-approval (testing only, not part of standard OAuth)
          example: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
      responses:
        "302":
          description: |
            Redirect to client with authorization code. The `Location` header contains the
            redirect URI with `code` and `state` query parameters.
          headers:
            Location:
              description: Redirect URI with authorization code
              schema:
                type: string
                example: "https://example.com/callback?code=SplxlOBeZQQYbYS6WxSbIA&state=abc123xyz"
        "400":
          description: Invalid request (missing parameters, invalid client, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
              example:
                error: invalid_request
                error_description: "client_id is required"
                state: "abc123xyz"
  /oauth/token:
    post:
      tags: [OAuth 2.0]
      summary: Token endpoint
      description: |
        Exchanges an authorization grant for tokens. Supports multiple grant types:

        - **authorization_code**: Exchange an authorization code for tokens (requires `code`, `redirect_uri`, optionally `code_verifier`)
        - **client_credentials**: Get a token for the client itself (no user involved)
        - **refresh_token**: Get a new access token using a refresh token
        - **device_code**: Poll for device authorization completion (RFC 8628)
        - **CIBA**: Poll for backchannel authentication completion

        Client can authenticate via HTTP Basic auth or `client_id`/`client_secret` in the body.
      operationId: token
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type:
                  type: string
                  enum:
                    - authorization_code
                    - client_credentials
                    - refresh_token
                    - urn:ietf:params:oauth:grant-type:device_code
                    - urn:openid:params:grant-type:ciba
                  description: The grant type to use
                code:
                  type: string
                  description: Authorization code received from /oauth/authorize (authorization_code grant)
                redirect_uri:
                  type: string
                  description: Must match the redirect_uri from the authorization request (authorization_code grant)
                code_verifier:
                  type: string
                  description: PKCE code verifier that corresponds to the code_challenge (authorization_code grant)
                refresh_token:
                  type: string
                  description: Refresh token to exchange for a new access token (refresh_token grant)
                device_code:
                  type: string
                  description: Device code received from /oauth/device (device_code grant)
                auth_req_id:
                  type: string
                  description: Auth request ID received from /oauth/bc-authorize (CIBA grant)
                scope:
                  type: string
                  description: Requested scopes (optional, may be reduced by server)
                resource:
                  type: string
                  description: |
                    Resource indicator (RFC 8707). The URI of the resource server.
                    The issued token will be scoped to this resource.
                  example: "https://api.example.com"
                client_id:
                  type: string
                  description: Client ID (alternative to HTTP Basic auth)
                client_secret:
                  type: string
                  description: Client secret (alternative to HTTP Basic auth)
            examples:
              authorization_code:
                summary: Authorization Code Grant with PKCE
                value:
                  grant_type: authorization_code
                  code: "SplxlOBeZQQYbYS6WxSbIA"
                  redirect_uri: "https://example.com/callback"
                  code_verifier: "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
              client_credentials:
                summary: Client Credentials Grant
                value:
                  grant_type: client_credentials
                  scope: "read write"
              refresh_token:
                summary: Refresh Token Grant
                value:
                  grant_type: refresh_token
                  refresh_token: "tGzv3JOkF0XG5Qx2TlKWIA"
              device_code:
                summary: Device Code Grant (polling)
                value:
                  grant_type: "urn:ietf:params:oauth:grant-type:device_code"
                  device_code: "GmRhmhcxhwAzkoEqiMEg_DnyEysNkuNhszIySk9eS"
              ciba:
                summary: CIBA Grant (polling)
                value:
                  grant_type: "urn:openid:params:grant-type:ciba"
                  auth_req_id: "1c266114-a1be-4252-8ad1-04986c5b9ac1"
      responses:
        "200":
          description: Token response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              example:
                access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type: "Bearer"
                expires_in: 3600
                refresh_token: "tGzv3JOkF0XG5Qx2TlKWIA"
                scope: "openid profile"
                id_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          description: Error response (invalid grant, expired code, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
              examples:
                invalid_grant:
                  summary: Invalid or expired authorization code
                  value:
                    error: invalid_grant
                    error_description: "Invalid authorization code"
                authorization_pending:
                  summary: Device/CIBA not yet authorized (continue polling)
                  value:
                    error: authorization_pending
                    error_description: "User has not yet approved the request"
                slow_down:
                  summary: Polling too fast (increase interval)
                  value:
                    error: slow_down
                    error_description: "Polling too frequently"
  /oauth/revoke:
    post:
      tags: [OAuth 2.0]
      summary: Token revocation (RFC 7009)
      description: |
        Revokes an access token or refresh token. Invalidates the token immediately.
        Revoking a refresh token also invalidates the associated access token.

        See [RFC 7009](https://datatracker.ietf.org/doc/html/rfc7009) for details.
      operationId: revokeToken
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: The token to revoke
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                  description: Hint about the token type (optional, helps server locate token faster)
                client_id:
                  type: string
                  description: Client ID (alternative to HTTP Basic auth)
                client_secret:
                  type: string
                  description: Client secret (alternative to HTTP Basic auth)
            example:
              token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
              token_type_hint: "access_token"
      responses:
        "200":
          description: Token revoked successfully (or token was already invalid)
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
  /oauth/introspect:
    post:
      tags: [OAuth 2.0]
      summary: Token introspection (RFC 7662)
      description: |
        Allows a protected resource to query the authorization server about the state of a token.
        Returns whether the token is active, along with metadata like scopes, client_id, and subject.

        See [RFC 7662](https://datatracker.ietf.org/doc/html/rfc7662) for details.
      operationId: introspectToken
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: The token to introspect
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                  description: Hint about the token type
                client_id:
                  type: string
                  description: Client ID (alternative to HTTP Basic auth)
                client_secret:
                  type: string
                  description: Client secret (alternative to HTTP Basic auth)
            example:
              token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        "200":
          description: Token introspection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrospectionResponse"
              examples:
                active_token:
                  summary: Active token
                  value:
                    active: true
                    scope: "openid profile"
                    client_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    username: "johndoe"
                    token_type: "Bearer"
                    exp: 1718003600
                    iat: 1718000000
                    sub: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
                inactive_token:
                  summary: Inactive (expired or revoked) token
                  value:
                    active: false
...
  /oidc/userinfo:
    get:
      tags: [OIDC]
      summary: UserInfo endpoint
      description: |
        Returns claims about the authenticated user. Requires a valid access token with `openid` scope.
        The claims returned depend on the scopes granted: `profile`, `email`, `phone`.
      operationId: userInfo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: UserInfo response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
              example:
                sub: "b5c6d7e8-f9a0-1234-5678-9abcdef01234"
                name: "johndoe"
                preferred_username: "johndoe"
                given_name: "John"
                family_name: "Doe"
                email: "john@example.com"
                email_verified: true
                phone_number: "+1-555-123-4567"
                phone_number_verified: true
                picture: "https://example.com/johndoe/photo.jpg"
                locale: "en-US"
                updated_at: 1718000000
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
              example:
                error: invalid_token
                error_description: "Token not found"

  # ──────────────────────────────────────────────
  # Management API Endpoints
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque OAuth 2.0 access token. Management routes require a client-credentials token whose scope contains management.
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic authentication with client_id:client_secret
...